tracking pins and start gui alterations
